refactor: add-harness flags mirror spec schema, remove toX mappers
