+++ src/handlers/project/add/harness/index.ts
@@ -3,27 +3,8 @@ import { createHandler, flag, ProjectKey } from "../../../../router";
 import type { AddProjectResourceConfig } from "../types";
 import { parseJsonFlag } from "../../../utils";
 import { InputValidationError } from "../../../../errors";
-import type {
-  AuthorizerConfiguration as SdkAuthorizerConfiguration,
-  HarnessEnvironmentArtifact,
-  HarnessEnvironmentProviderRequest,
-  HarnessGatewayOutboundAuth as SdkHarnessGatewayOutboundAuth,
-  HarnessMemoryConfiguration as SdkMemoryConfiguration,
-  HarnessModelConfiguration,
-  HarnessSkill as SdkHarnessSkill,
-  HarnessTool as SdkHarnessTool,
-  HarnessTruncationConfiguration as SdkTruncationConfiguration,
-} from "@aws-sdk/client-bedrock-agentcore-control";
-import type {
-  HarnessGatewayOutboundAuth,
-  HarnessMemoryRef,
-  HarnessModel,
-  HarnessSkill,
-  HarnessTool,
-  HarnessTruncationConfig,
-  ManagedMemoryStrategy,
-} from "../../../../projectSchemas/harness";
-import type { AuthorizerConfig } from "../../../../projectSchemas/auth";
+import { HarnessSpecSchema } from "../../../../projectSchemas/harness";
+import { formatZodError } from "../../../../router/schema";
 
 export const createAddHarnessHandler = (config: AddProjectResourceConfig) =>
   createHandler({
@@ -37,42 +18,40 @@ export const createAddHarnessHandler = (config: AddProjectResourceConfig) =>
         z.string().optional(),
       ),
       flag("system-prompt", "the agent's system prompt", z.string().optional()),
-      flag("model", "model configuration (JSON HarnessModelConfiguration)", z.string().optional()),
-      flag("tools", "tools available to the agent (JSON HarnessTool[])", z.string().optional()),
-      flag("skills", "skills available to the agent (JSON HarnessSkill[])", z.string().optional()),
+      flag("model", "model configuration (JSON)", z.string().optional()),
+      flag("tools", "tools available to the agent (JSON)", z.string().optional()),
+      flag("skills", "skills available to the agent (JSON)", z.string().optional()),
       flag(
         "allowed-tools",
         "tool allowlist patterns (e.g. * or @serverName/toolName)",
         z.array(z.string()).optional(),
       ),
+      flag("memory", "memory configuration (JSON)", z.string().optional()),
+      flag("truncation", "context truncation configuration (JSON)", z.string().optional()),
       flag(
-        "memory",
-        "memory configuration (JSON HarnessMemoryConfiguration)",
+        "network-mode",
+        "network mode for the harness environment (PUBLIC or VPC)",
         z.string().optional(),
       ),
-      flag(
-        "truncation",
-        "context truncation configuration (JSON HarnessTruncationConfiguration)",
-        z.string().optional(),
-      ),
-      flag(
-        "environment",
-        "compute environment configuration (JSON HarnessEnvironmentProviderRequest)",
-        z.string().optional(),
-      ),
+      flag("network-config", "VPC network configuration (JSON)", z.string().optional()),
+      flag("lifecycle-config", "lifecycle configuration (JSON)", z.string().optional()),
+      flag("session-storage-path", "mount path for session storage", z.string().optional()),
+      flag("efs-access-points", "EFS access point configurations (JSON)", z.string().optional()),
+      flag("s3-access-points", "S3 access point configurations (JSON)", z.string().optional()),
       flag(
         "environment-variables",
         "environment variables (JSON object of key/value strings)",
         z.string().optional(),
       ),
+      flag("container-uri", "ECR container image URI", z.string().optional()),
       flag(
-        "environment-artifact",
-        "environment artifact configuration (ex. container image) (JSON HarnessEnvironmentArtifact)",
+        "authorizer-type",
+        "inbound authorizer type (AWS_IAM or CUSTOM_JWT)",
         z.string().optional(),
       ),
       flag(
         "authorizer-configuration",
-        "inbound authorizer configuration (JSON AuthorizerConfiguration)",
+        "inbound authorizer configuration (JSON)",
         z.string().optional(),
       ),
       flag("max-iterations", "max agent loop iterations per invocation", z.number().optional()),
@@ -84,88 +63,56 @@ export const createAddHarnessHandler = (config: AddProjectResourceConfig) =>
         "path to local dockerfile to use as the container image for the harness",
         z.string().optional(),
       ),
-      flag(
-        "vpc-id",
-        "VPC ID for Dockerfile builds in VPC mode (required when combining --dockerfile with VPC networking)",
-        z.string().optional(),
-      ),
     ],
     handle: async (ctx, flags) => {
       if (!flags.name)
         throw new InputValidationError("required option '--name <name>' not specified");
 
-      const inputModelConfig = parseJsonFlag<HarnessModelConfiguration>("model", flags["model"]);
-      const inputTools = parseJsonFlag<SdkHarnessTool[]>("tools", flags["tools"]);
-      const inputSkills = parseJsonFlag<SdkHarnessSkill[]>("skills", flags["skills"]);
-      const inputMemory = parseJsonFlag<SdkMemoryConfiguration>("memory", flags["memory"]);
-      const inputTruncation = parseJsonFlag<SdkTruncationConfiguration>(
-        "truncation",
-        flags["truncation"],
-      );
-      const inputAuthConfig = parseJsonFlag<SdkAuthorizerConfiguration>(
-        "authorizer-configuration",
-        flags["authorizer-configuration"],
-      );
-      const inputEnvironment = parseJsonFlag<HarnessEnvironmentProviderRequest>(
-        "environment",
-        flags["environment"],
-      );
-      const inputArtifact = parseJsonFlag<HarnessEnvironmentArtifact>(
-        "environment-artifact",
-        flags["environment-artifact"],
-      );
-      const env = inputEnvironment ? toEnvironment(inputEnvironment) : undefined;
-      const artifact = inputArtifact ? toEnvironmentArtifact(inputArtifact) : undefined;
-
-      const inputVpcId = flags["vpc-id"];
-
-      if (inputArtifact?.containerConfiguration?.containerUri && flags.dockerfile)
-        throw new InputValidationError(`containerUri and dockerfile are mutually exclusive`);
-
-      if (inputVpcId && !env?.networkConfig) {
-        throw new InputValidationError(
-          "--vpc-id requires --environment with VPC network configuration",
-        );
-      }
-
-      const harnessConfig = {
+      const harnessInput = {
         name: flags.name,
-        model: inputModelConfig
-          ? toModelConfig(inputModelConfig)
-          : { provider: "bedrock" as const, modelId: "global.anthropic.claude-sonnet-4-6" },
+        model: parseJsonFlag("model", flags["model"]) ?? {
+          provider: "bedrock" as const,
+          modelId: "global.anthropic.claude-sonnet-4-6",
+        },
         systemPrompt: flags["system-prompt"],
         executionRoleArn: flags["execution-role-arn"],
-        tools: inputTools?.map(toTool),
-        skills: inputSkills?.map(toSkill),
+        tools: parseJsonFlag("tools", flags["tools"]),
+        skills: parseJsonFlag("skills", flags["skills"]),
         allowedTools: flags["allowed-tools"],
-        memory: inputMemory ? toMemory(inputMemory) : undefined,
-        truncation: inputTruncation ? toTruncation(inputTruncation) : undefined,
-        environmentVariables: parseJsonFlag<Record<string, string>>(
+        memory: parseJsonFlag("memory", flags["memory"]),
+        truncation: parseJsonFlag("truncation", flags["truncation"]),
+        networkMode: flags["network-mode"],
+        networkConfig: parseJsonFlag("network-config", flags["network-config"]),
+        lifecycleConfig: parseJsonFlag("lifecycle-config", flags["lifecycle-config"]),
+        sessionStoragePath: flags["session-storage-path"],
+        efsAccessPoints: parseJsonFlag("efs-access-points", flags["efs-access-points"]),
+        s3AccessPoints: parseJsonFlag("s3-access-points", flags["s3-access-points"]),
+        environmentVariables: parseJsonFlag(
           "environment-variables",
           flags["environment-variables"],
         ),
-        authorizerType: inputAuthConfig ? ("CUSTOM_JWT" as const) : undefined,
-        authorizerConfiguration: inputAuthConfig ? toAuthorizerConfig(inputAuthConfig) : undefined,
+        containerUri: flags["container-uri"],
+        authorizerType: flags["authorizer-type"],
+        authorizerConfiguration: parseJsonFlag(
+          "authorizer-configuration",
+          flags["authorizer-configuration"],
+        ),
         maxIterations: flags["max-iterations"],
         maxTokens: flags["max-tokens"],
         timeoutSeconds: flags["timeout-seconds"],
-        tags: parseJsonFlag<Record<string, string>>("tags", flags["tags"]),
-        networkMode: env?.networkMode,
-        networkConfig: env?.networkConfig
-          ? { ...env.networkConfig, ...(inputVpcId ? { vpcId: inputVpcId } : {}) }
-          : undefined,
-        lifecycleConfig: env?.lifecycleConfig,
-        sessionStoragePath: env?.sessionStoragePath,
-        efsAccessPoints: env?.efsAccessPoints,
-        s3AccessPoints: env?.s3AccessPoints,
-        containerUri: artifact?.containerUri,
+        tags: parseJsonFlag("tags", flags["tags"]),
         dockerfile: flags["dockerfile"],
       };
+
+      const result = HarnessSpecSchema.safeParse(harnessInput);
+      if (!result.success) {
+        throw new InputValidationError(formatZodError(result.error), { cause: result.error });
+      }
 
       const project = ctx.require(ProjectKey);
       for await (const event of config.projectManager.addResource(project, {
         resourceType: "harness",
-        resourceConfig: harnessConfig,
+        resourceConfig: result.data,
       })) {
         config.io.stderr.write(`${event.message}\n`);
       }
@@ -173,317 +120,3 @@ export const createAddHarnessHandler = (config: AddProjectResourceConfig) =>
       config.io.stderr.write(`added harness '${flags["name"]}' to '${project.name}'\n`);
     },
   });
-
-/** Converts the SDK's tagged-union model config into the flat project-schema shape. */
-function toModelConfig(modelConfig: HarnessModelConfiguration): HarnessModel {
-  function commonFields(c: {
-    modelId?: string;
-    maxTokens?: number;
-    temperature?: number;
-    topP?: number;
-    additionalParams?: unknown;
-  }) {
-    if (!c.modelId) throw new InputValidationError("modelId is required in model configuration");
-    return {
-      modelId: c.modelId,
-      maxTokens: c.maxTokens,
-      temperature: c.temperature,
-      topP: c.topP,
-      additionalParams: c.additionalParams as Record<string, unknown> | undefined,
-    };
-  }
-
-  if ("bedrockModelConfig" in modelConfig && modelConfig.bedrockModelConfig) {
-    const c = modelConfig.bedrockModelConfig;
-    return { provider: "bedrock", ...commonFields(c), apiFormat: c.apiFormat };
-  }
-  if ("openAiModelConfig" in modelConfig && modelConfig.openAiModelConfig) {
-    const c = modelConfig.openAiModelConfig;
-    return {
-      provider: "open_ai",
-      ...commonFields(c),
-      apiKeyArn: c.apiKeyArn,
-      apiFormat: c.apiFormat,
-    };
-  }
-  if ("geminiModelConfig" in modelConfig && modelConfig.geminiModelConfig) {
-    const c = modelConfig.geminiModelConfig;
-    return { provider: "gemini", ...commonFields(c), apiKeyArn: c.apiKeyArn, topK: c.topK };
-  }
-  if ("liteLlmModelConfig" in modelConfig && modelConfig.liteLlmModelConfig) {
-    const c = modelConfig.liteLlmModelConfig;
-    return { provider: "lite_llm", ...commonFields(c), apiKeyArn: c.apiKeyArn, apiBase: c.apiBase };
-  }
-  throw new InputValidationError("Unrecognized model configuration variant");
-}
-
-/** Converts an SDK HarnessTool into the flat project-schema shape. */
-function toTool(tool: SdkHarnessTool): HarnessTool {
-  if (!tool.type) throw new InputValidationError("tool type is required");
-  if (!tool.name) throw new InputValidationError(`tool name is required (type: ${tool.type})`);
-  if (!tool.config) return { type: tool.type, name: tool.name };
-  const c = tool.config;
-  if ("remoteMcp" in c && c.remoteMcp) {
-    return {
-      type: tool.type,
-      name: tool.name,
-      config: {
-        remoteMcp: {
-          url: requireField(c.remoteMcp.url, "remoteMcp.url"),
-          headers: c.remoteMcp.headers,
-        },
-      },
-    };
-  }
-  if ("agentCoreBrowser" in c && c.agentCoreBrowser) {
-    return {
-      type: tool.type,
-      name: tool.name,
-      config: { agentCoreBrowser: { browserArn: c.agentCoreBrowser.browserArn } },
-    };
-  }
-  if ("agentCoreGateway" in c && c.agentCoreGateway) {
-    return {
-      type: tool.type,
-      name: tool.name,
-      config: {
-        agentCoreGateway: {
-          gatewayArn: requireField(c.agentCoreGateway.gatewayArn, "agentCoreGateway.gatewayArn"),
-          outboundAuth: c.agentCoreGateway.outboundAuth
-            ? toOutboundAuth(c.agentCoreGateway.outboundAuth)
-            : undefined,
-        },
-      },
-    };
-  }
-  if ("inlineFunction" in c && c.inlineFunction) {
-    return {
-      type: tool.type,
-      name: tool.name,
-      config: {
-        inlineFunction: {
-          description: requireField(c.inlineFunction.description, "inlineFunction.description"),
-          inputSchema: c.inlineFunction.inputSchema as Record<string, unknown>,
-        },
-      },
-    };
-  }
-  if ("agentCoreCodeInterpreter" in c && c.agentCoreCodeInterpreter) {
-    return {
-      type: tool.type,
-      name: tool.name,
-      config: {
-        agentCoreCodeInterpreter: {
-          codeInterpreterArn: c.agentCoreCodeInterpreter.codeInterpreterArn,
-        },
-      },
-    };
-  }
-  return { type: tool.type, name: tool.name };
-}
-
-/** Converts an SDK HarnessGatewayOutboundAuth tagged union into the project-schema shape. */
-function toOutboundAuth(auth: SdkHarnessGatewayOutboundAuth): HarnessGatewayOutboundAuth {
-  if ("awsIam" in auth && auth.awsIam) return { awsIam: {} };
-  if ("none" in auth && auth.none) return { none: {} };
-  if ("oauth" in auth && auth.oauth) {
-    return {
-      oauth: {
-        providerArn: requireField(auth.oauth.providerArn, "outboundAuth.oauth.providerArn"),
-        scopes: requireField(auth.oauth.scopes, "outboundAuth.oauth.scopes"),
-        // SDK does not expose this type directly.
-        grantType: auth.oauth.grantType as Extract<
-          HarnessGatewayOutboundAuth,
-          { oauth: unknown }
-        >["oauth"]["grantType"],
-        customParameters: auth.oauth.customParameters,
-      },
-    };
-  }
-  throw new InputValidationError("unrecognized outboundAuth variant");
-}
-
-/** Converts an SDK HarnessSkill tagged union into the project-schema shape. */
-function toSkill(skill: SdkHarnessSkill): HarnessSkill {
-  if ("path" in skill && skill.path) {
-    return { path: skill.path };
-  }
-  if ("s3" in skill && skill.s3) {
-    return { s3Uri: requireField(skill.s3.uri, "skill.s3.uri") };
-  }
-  if ("git" in skill && skill.git) {
-    return {
-      gitUrl: requireField(skill.git.url, "skill.git.url"),
-      path: skill.git.path,
-      auth: skill.git.auth
-        ? {
-            credentialArn: requireField(
-              skill.git.auth.credentialArn,
-              "skill.git.auth.credentialArn",
-            ),
-            username: skill.git.auth.username,
-          }
-        : undefined,
-    };
-  }
-  if ("awsSkills" in skill && skill.awsSkills) {
-    return { awsSkills: { paths: skill.awsSkills.paths } };
-  }
-  throw new InputValidationError("Unrecognized skill variant");
-}
-
-/** Converts an SDK HarnessMemoryConfiguration tagged union into the project-schema shape. */
-function toMemory(memory: SdkMemoryConfiguration): HarnessMemoryRef {
-  if ("managedMemoryConfiguration" in memory && memory.managedMemoryConfiguration) {
-    const c = memory.managedMemoryConfiguration;
-    return {
-      mode: "managed",
-      strategies: c.strategies as ManagedMemoryStrategy[] | undefined,
-      eventExpiryDuration: c.eventExpiryDuration,
-      encryptionKeyArn: c.encryptionKeyArn,
-    };
-  }
-  if ("agentCoreMemoryConfiguration" in memory && memory.agentCoreMemoryConfiguration) {
-    const c = memory.agentCoreMemoryConfiguration;
-    return {
-      mode: "existing",
-      arn: c.arn,
-      actorId: c.actorId,
-      messagesCount: c.messagesCount,
-    };
-  }
-  if ("disabled" in memory && memory.disabled) {
-    return { mode: "disabled" };
-  }
-  throw new InputValidationError("Unrecognized memory configuration variant");
-}
-
-/** Converts an SDK HarnessTruncationConfiguration into the project-schema shape. */
-function toTruncation(truncation: SdkTruncationConfiguration): HarnessTruncationConfig {
-  if (!truncation.strategy) throw new InputValidationError("truncation strategy is required");
-  const config = truncation.config;
-  if (!config) return { strategy: truncation.strategy };
-  if ("slidingWindow" in config && config.slidingWindow) {
-    return {
-      strategy: truncation.strategy,
-      config: { slidingWindow: { messagesCount: config.slidingWindow.messagesCount } },
-    };
-  }
-  if ("summarization" in config && config.summarization) {
-    return {
-      strategy: truncation.strategy,
-      config: {
-        summarization: {
-          summaryRatio: config.summarization.summaryRatio,
-          preserveRecentMessages: config.summarization.preserveRecentMessages,
-          summarizationSystemPrompt: config.summarization.summarizationSystemPrompt,
-        },
-      },
-    };
-  }
-  return { strategy: truncation.strategy };
-}
-
-/** Converts an SDK AuthorizerConfiguration tagged union into the project-schema shape. */
-function toAuthorizerConfig(auth: SdkAuthorizerConfiguration): AuthorizerConfig {
-  if ("customJWTAuthorizer" in auth && auth.customJWTAuthorizer) {
-    const c = auth.customJWTAuthorizer;
-    if (!c.discoveryUrl)
-      throw new InputValidationError("discoveryUrl is required in authorizer configuration");
-    return {
-      customJwtAuthorizer: {
-        discoveryUrl: c.discoveryUrl,
-        allowedAudience: c.allowedAudience,
-        allowedClients: c.allowedClients,
-        allowedScopes: c.allowedScopes,
-      },
-    };
-  }
-  throw new InputValidationError("Unrecognized authorizer configuration variant");
-}
-
-/** Decomposes the SDK's environment tagged union into flat HarnessSpec fields. */
-function toEnvironment(env: HarnessEnvironmentProviderRequest) {
-  if (!("agentCoreRuntimeEnvironment" in env) || !env.agentCoreRuntimeEnvironment) {
-    throw new InputValidationError("Unrecognized environment configuration variant");
-  }
-  const rt = env.agentCoreRuntimeEnvironment;
-  const net = rt.networkConfiguration;
-  const fss = rt.filesystemConfigurations ?? [];
-
-  const sessionStorage = fss.find((f) => "sessionStorage" in f && f.sessionStorage);
-  const efsAccessPoints = fss.filter((f) => "efsAccessPoint" in f && f.efsAccessPoint);
-  const s3AccessPoints = fss.filter((f) => "s3FilesAccessPoint" in f && f.s3FilesAccessPoint);
-
-  return {
-    networkMode: net?.networkMode as "PUBLIC" | "VPC" | undefined,
-    networkConfig: net?.networkModeConfig
-      ? {
-          subnets: requireField(net.networkModeConfig.subnets, "networkConfiguration.subnets"),
-          securityGroups: requireField(
-            net.networkModeConfig.securityGroups,
-            "networkConfiguration.securityGroups",
-          ),
-        }
-      : undefined,
-    lifecycleConfig: rt.lifecycleConfiguration
-      ? {
-          idleRuntimeSessionTimeout: rt.lifecycleConfiguration.idleRuntimeSessionTimeout,
-          maxLifetime: rt.lifecycleConfiguration.maxLifetime,
-        }
-      : undefined,
-    sessionStoragePath:
-      sessionStorage && "sessionStorage" in sessionStorage
-        ? requireField(
-            ("sessionStorage" in sessionStorage ? sessionStorage.sessionStorage : undefined)
-              ?.mountPath,
-            "sessionStorage.mountPath",
-          )
-        : undefined,
-    efsAccessPoints:
-      efsAccessPoints.length > 0
-        ? efsAccessPoints.map((f) => {
-            const efs = requireField(
-              "efsAccessPoint" in f ? f.efsAccessPoint : undefined,
-              "efsAccessPoint",
-            );
-            return {
-              accessPointArn: requireField(efs.accessPointArn, "efsAccessPoint.accessPointArn"),
-              mountPath: requireField(efs.mountPath, "efsAccessPoint.mountPath"),
-            };
-          })
-        : undefined,
-    s3AccessPoints:
-      s3AccessPoints.length > 0
-        ? s3AccessPoints.map((f) => {
-            const s3 = requireField(
-              "s3FilesAccessPoint" in f ? f.s3FilesAccessPoint : undefined,
-              "s3FilesAccessPoint",
-            );
-            return {
-              accessPointArn: requireField(s3.accessPointArn, "s3FilesAccessPoint.accessPointArn"),
-              mountPath: requireField(s3.mountPath, "s3FilesAccessPoint.mountPath"),
-            };
-          })
-        : undefined,
-  };
-}
-
-/** Decomposes the SDK's environment artifact tagged union into flat HarnessSpec fields. */
-function toEnvironmentArtifact(artifact: HarnessEnvironmentArtifact) {
-  if ("containerConfiguration" in artifact && artifact.containerConfiguration) {
-    return {
-      containerUri: requireField(
-        artifact.containerConfiguration.containerUri,
-        "containerConfiguration.containerUri",
-      ),
-    };
-  }
-  throw new InputValidationError("Unrecognized environment artifact variant");
-}
-
-/** Validates a required field is present, throwing with context instead of crashing opaquely. */
-function requireField<T>(value: T | undefined | null, field: string): T {
-  if (value == null) throw new InputValidationError(`${field} is required`);
-  return value;
-}
